backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
botocore==1.34.0
cryptography
pydantic[email]==2.5.0
python-dotenv==1.0.0
PyJWT